Move ipywidgets as well
