+++ src/test/datascience/ipywidgets/cdnWidgetScriptSourceProvider.unit.test.ts
@@ -15,10 +15,10 @@ import { IFileSystem } from '../../../client/common/platform/types';
 import { IConfigurationService, WidgetCDNs } from '../../../client/common/types';
 import { noop } from '../../../client/common/utils/misc';
 import { EXTENSION_ROOT_DIR } from '../../../client/constants';
-import { CDNWidgetScriptSourceProvider } from '../../../client/datascience/ipywidgets/cdnWidgetScriptSourceProvider';
-import { IPyWidgetScriptSource } from '../../../client/datascience/ipywidgets/ipyWidgetScriptSource';
-import { IWidgetScriptSourceProvider } from '../../../client/datascience/ipywidgets/types';
 import { ILocalResourceUriConverter } from '../../../client/datascience/types';
+import { CDNWidgetScriptSourceProvider } from '../../../kernels/ipywidgets/cdnWidgetScriptSourceProvider';
+import { IPyWidgetScriptSource } from '../../../kernels/ipywidgets/ipyWidgetScriptSource';
+import { IWidgetScriptSourceProvider } from '../../../kernels/ipywidgets/types';
 
 /* eslint-disable @typescript-eslint/no-var-requires, @typescript-eslint/no-require-imports, , @typescript-eslint/no-explicit-any, , no-console */
 const sanitize = require('sanitize-filename');
