+++ src/test/datascience/notebook/ipywidget.vscode.test.ts
@@ -24,8 +24,8 @@ import {
 import { EXTENSION_ROOT_DIR_FOR_TESTS } from '../../constants';
 import { createDeferred, Deferred } from '../../../client/common/utils/async';
 import { InteractiveWindowMessages } from '../../../client/datascience/interactive-common/interactiveWindowTypes';
-import { NotebookIPyWidgetCoordinator } from '../../../client/datascience/ipywidgets/notebookIPyWidgetCoordinator';
 import { INotebookCommunication } from '../../../client/datascience/notebook/types';
+import { NotebookIPyWidgetCoordinator } from '../../../kernels/ipywidgets/notebookIPyWidgetCoordinator';
 
 /* eslint-disable @typescript-eslint/no-explicit-any, no-invalid-this */
 suite('DataScience - VSCode Notebook - IPyWidget test', () => {
